removed last traces of boto 2 from mrjob/fs/s3.py
